Fixed checkstyle, added missing javadocs and readme
